Rename session manager and create global renderer
